quizscreen setup in screens
